[Button] Reintegrate propTypes and eventHandlers
